[Containers] Separate concept of "List Items" from "Container Items" #1271 (#2416)
* Improve panel container tests

Updates panel container tests to check
the panel data layers, as well as the
panel IDs and paths.

* Rename `PanelContainerItemImpl` to `PanelContainerListItemImpl`

Renames the `PanelContainerItemImpl` to clarify that this is
a ListItem implementation.

* Rename `PanelContainerImpl` -> `AbstractPanelContainerImpl`

The `PanelContainerImpl` is not a concrete class, it is abstract.
The class is renamed to follow the naming convention of other
abstract classes in the package.

* Deprecate `PanelContainerListItemImpl`

This `ListItem` implementation is only used by panel container
components. Panel container components will now prefer to return
a `PanelContainerItem`, which gives direct access to the resource,
as well as any other panel metadata that is needed.

* Deprecate `ResourceListItemImpl`

This `ListItem` implementation is only used by container
components. Container components will now prefer to return a
`ContainerItem`, which gives direct access to the resource.

* Add utility method for getting child components and component exporters

Adds general utility methods to the `ComponentUtils` class for the purpose
of finding child resources that are components, and child resources that
are component exporters.

This logic is moved to the utility class because it is likely to be reused
by numerous container components, and is likely to be useful for custom
client container models.

* Convert `AbstractContainerImpl` to use `ComponentUtils`

Updates the `Container` model implementations to use the newly
added helper methods`ComponentUtils#getComponentModels()` and
`ComponentUtils#getChildComponents()`.

The return type of `AbstractContainerImpl#getExportedItems()`
is changed to `LinkedHashMap` to indicate the ordering of these
items is critically important.

The `AbstractContainerImpl#getExportedItems()` method is made
abstract, instead of relying on other abstract methods to
provide the data. This clarifies the application flow without
introducing undue duplication of logic.

* Move `AbstractContainer#getChildren()` method to implementations

Moves the `#getChildren()` method in the `AbstractContainer` class
deeper in the class hierarchy to `LayoutContainerImpl` and
`PanelContainerImpl`.

Currently, the method is duplicated in these locations, but will
diverge in the future.

* Deprecate `Container#getItems()` and `AbstractContainerImpl#readItems()`

The `readItems()` abstract method was a means by which to delegate the
constructing of container `ListItem`s to the component model.
The `ListItems()` method was the mechanism by which those `ListItem`s
were exposed.

It is no longer desired that containers should be treated as lists;
therefore, these methods are marked as deprecated.

* Add `ContainerItem` and `PanelContainerItem`

Adds interfaces and implementations for `ContainerItem` and
`PanelContainerItem`. These are the new non-ListItem classes
that should be offered by `Container` and `PanelContainer`
components, respectively.

* Update container models to return `ContainerItem`s

Adds the `#getChildren()` method to the `Container` interface
and implementations.

Adds the `PanelContainer` interface, which is a specialized
case of containers, where the children are `PanelContainerItem`s.

* Bump `com.adobe.cq.wcm.core.components.util` version -> 1.1.0

Updates the version of package `com.adobe.cq.wcm.core.components.util`
to 1.1.0

* Bump `com.adobe.cq.wcm.core.components.models` version -> 12.19.0

Updates the version of package `com.adobe.cq.wcm.core.components.models`
to 12.19.0.

* Update container component HTL to use `ContainerItem`

Updates the container components HTL to use `ContainerItem`s
instead of `ListItem`s.

* Default getChildren

- provide default implementation for getChildren to prevent breaking change for customer

* Revert "Default getChildren"

This reverts commit f385dd52b0162435acedc1f00af1a65befe1b9bc.

* Revert ListItem#getResource

- revert ListItem#getResource
- make simple.html backward compatible

* Revert ListItem#getResource

- fixed removed PanelContainerImpl RT

* Merge branch 'main' into issue/1031-container-item

- fix changed constant field after merge

* review: removed jcr:lastModified in test content

* review: removed jcr:lastModified in test content

* Added item.getName and implemented fallback for getItems in case getChildren is not implemented

* fixed issue in tabs.html

---------

Co-authored-by: Kyle Giovannetti <kyle.giovannetti@dal.ca>
Co-authored-by: Dirk Rudolph <drudolph@adobe.com>
Co-authored-by: Vlad Bailescu <vlad@bailescu.ro>