Change `Utils#getSuperTypes` to use `ResourceResolver#getParentResourceType(String);
Updates `Utils#getSuperType` to use `resourceResolver.getParentResourceType(superType)`
instead of `resourceResolver.getResource(superType).getResourceSuperType`.

This change is made because any `ResourceResolver` can always resolve the resource
super type hierarchy via `ResourceResolver#getParentResourceType(String)`, even if
`ResrouceResolver#getResource(String)` would be unable to resolve the super type
resource due to permissions.

After this change, the only requirement for the `ResourceResolver` passed to
`Utils#getSuperTypes()` is that it be an active `ResourceResolver` - it does
not need to be a special `ResourceResolver` with any elevated permissions.

----
refs issue #2446