Use empty list supplier for `BreadcrumbItem` children.
Updates the breadcrumb model implementations by removing the `children` field
from `BreadcrumbItemImpl` constructor because it is always set to empty list (i.e.
breadcrumb items never have children).

The `BreadcrumbItemImpl` will now pass a `Supplier` which
supplies an empty list to the super class `NavigationItemImpl`.

Removes `BreadcrumbImpl#getExportedType()` because it is exactly the same
as the super method.

----
refs #2526