Update V1/V2 `LanguageNavigationImpl` to pass `Supplier` to item constructors
Updates the V1 and V2 `LanguageNavigationImpl` models to pass
a `Supplier` to their respective `LanguageNavigationItemImpl`
constructors

Fixes a possible NPE in `LanguageNavigationImpl` if
navigation root is not set in either properties or
policy.

Removes `LanguageNavigationImpl#getExportedType()` because it
is exactly the same as the method from the super class.

Update how titles are determined for language navigation items to
reuse the logic used for all other navigation items.

----
refs #2526