[SITES-13898] Fixed moved item deletion for the Accordion/Carousel/Tabs components (#2547)
Fixed an issue that occurred when we tried to delete an item that was moved before the deletion - in that case it wouldn't get deleted.
The problem is that the move operation is effectively a combination of the two operations - delete and add.
The way we handled it before was that we would store the name of the item that triggered "coral-multifield:itemorder" and if later the "coral-collection:remove" is triggered for that same item - we would not add it to the deleted list since the item is considered to be moved. But if you would combine move and delete operations for the same item - we would do the move but not the delete since the delete operation will be skipped by the move.
Now we are always adding an item to the delete list if "coral-collection:remove" was triggered and in case of "coral-collection:add" we are making sure that the item that was just added will be removed from the deleted list.