SITES-30116 AEM Teaser/Image Core Component Alternative Text Field Issue (#2928)
Fixes alt text from page image not being shown inside image / teaser components on initial load.

Steps to reproduce:

Add a page image with alt text to a page inside WKND content, such as /content/wknd/es/es
Drag and drop an image or teaser component.
Select the "Inherit alternative text from page" checkbox in the Asset tab.
Observe the alternative text field, which should be populated but is instead disabled and blank.
Uncheck and then recheck the "Inherit alternative text from page" checkbox to see if the field populates correctly.
In the usual case it does not populate directly, likely due to a race condition because in certain cases, such as when console / debugger is open, the alt text is in fact populated correctly.

Issue is fixed by explicitly adding the alt text from page to the text field if it exists and the inherit alt text from page image checkbox is checked. The fix applies to both Image and Teaser Components since they use the same code here.

Additionally slightly cleaned up the code to use preexisting vars for selectors in a few places