ASSETS-46423 Support authoring preview versions of assets in sites core component
Add a preview token builder which, on getting a preview key in environment variables, will always generate and append the preview token to DM with OpenAPIs images. This preview token will ensure that preview version of an asset is visible in image component if it exists. For customers who do not have the preview secret key set in their environment variables, the existing delivery URL will be generated.